fixed stupid type error

+++ src/base.rs
@@ -1739,7 +1739,7 @@ impl MotifSetDef {
         let mut scrambled_motifs = false;
         for mot in set.iter_mut() {
             let same_sd = (mot.kernel.get_sd() - self.width/WIDTH_TO_SD).abs() < 1e-6;
-            let same_len = (WIDE*mot.kernel.get_sd()/data.spacer() - (mot.kernel.len() as f64)).abs() < 1e-6;
+            let same_len = (WIDE*mot.kernel.get_sd()/(data.spacer() as f64) - (mot.kernel.len() as f64)).abs() < 1e-6;
 
             if !(same_sd && same_len) {
                 changed_kernels = true;
